首次提交，涵盖了legato原始代码适配ubuntu24.04编译通过做的所有修改。

+++ apps/test/framework/inspect/mutexFlux/mutexFlux.c
@@ -204,7 +204,7 @@ void createAllMutexes
     long threadCnt = 0;
     while (threadCnt < ThreadNum)
     {
-        snprintf(threadNameBuffer, MAX_THREAD_NAME_SIZE, "Thread%ld", threadCnt);
+        snprintf(threadNameBuffer, MAX_THREAD_NAME_SIZE, "Thread%d", (int)threadCnt);
 
         // Spread mutexes evenly among the threads, and put the remaining mutexes in the last thread.
         mutexPerThread = (threadCnt == (ThreadNum - 1)) ? (quotient + remainder) : quotient;
